+++ crates/web-client/src/models/mod.rs
@@ -92,5 +92,4 @@ pub mod transaction_result;
 pub mod transaction_script;
 pub mod transaction_script_inputs;
 pub mod transaction_status;
-pub mod transactions;
 pub mod word;